updated loadunload ui while gaining resources
Former-commit-id: 6548ba88308c02d3dd288b0292a25b53e6143a7f
Former-commit-id: d2d11eaa2e630a351e2dfdff2349eb063992f6d7

+++ Assets/Scripts/UI/UIResourceManager.cs
@@ -146,7 +146,7 @@ public class UIResourceManager : MonoBehaviour
             buttonDown.sprite = buttonUp;
             overflowGridHolder.gameObject.SetActive(v);
             overflowActiveStatus = true;
-            float size = Mathf.CeilToInt((activeCells - gridWidth) / gridWidth) * 90;
+            float size = Mathf.CeilToInt((activeCells - gridWidth) / (float)gridWidth) * 90;
 
             overflowGridHolder.anchoredPosition3D = overflowOriginalLoc + new Vector3(0, size, 0);
 
@@ -156,7 +156,7 @@ public class UIResourceManager : MonoBehaviour
         {
             buttonDown.sprite = originalDown;
             overflowActiveStatus = false;
-            LeanTween.moveY(overflowGridHolder, overflowGridHolder.anchoredPosition3D.y + Mathf.CeilToInt((activeCells - gridWidth) / gridWidth) * 90, 0.2f).setOnComplete(SetOverflowStatusFalse);
+            LeanTween.moveY(overflowGridHolder, overflowGridHolder.anchoredPosition3D.y + Mathf.CeilToInt((activeCells - gridWidth) / (float)gridWidth) * 90, 0.2f).setOnComplete(SetOverflowStatusFalse);
         }
     }
